Tests for adding person, creating room and other things
Application files are in the app folder, which has two folders in it so far.
Folder Person contains the classes to do with People that is Person, Staff and Fellow
Person tests are included in this file
Folder Room contains the classes to do with Rooms, most of which still under progress.
Tests to do with the room are in this folder.
This commit also inlcudes a requirements file listing the libraries installed in the virtual environment
The libraries so far include docopt and nose for carrying out nose tests.
The designs folder is to have UML class designs and other diagrams yet to add