Random Allocation of Rooms
In this commit the method to randomly allocate rooms is done and passes it's tests.
The method is broken down to fisrt running a method to create person.
The create person method passed it's individual tests. Now since this method
is called within the project code, the risk of error in inputs is minimal.
After a method that assigns roooms is called, it has also passed it's tests